queue_services/strr-email/pyproject.toml
[tool.poetry]
name = "strr-email"
version = "1.1.10"
description = "BC Registries - strr-email"
authors = []
license = "BSD-3-Clause"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12.2"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.4"
launchdarkly-server-sdk = "^9.0.1"
python-dotenv = "^1.0.1"
sentry-sdk = "^1.44.1"
flask-babel = "^4.0.0"
psycopg2 = "^2.9.9"
flask-jwt-oidc = "^0.8.0"
gunicorn = "^21.2.0"
flask-cors = "^4.0.0"
pg8000 = "^1.31.2"
protobuf = "3.19.5"
isort = "^5.13.2"
pylint = "^3.0.3"
gcp-queue = { git = "https://github.com/bcgov/sbc-connect-common.git", subdirectory = "python/gcp-queue", branch = "main" }
structured-logging = { git = "https://github.com/bcgov/sbc-connect-common.git", subdirectory = "python/structured-logging", branch = "main" }
strr-api = {git = "https://github.com/bcgov/STRR.git", rev = "main", subdirectory = "strr-api"}
nanoid = "^2.0.0"
# strr-api = {path = "/workspaces/STRR/strr-api", develop = true}


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-mock = "^3.14.0"
black = "^24.4.2"
zimports = "^0.6.1"
isort = "^5.13.2"
testcontainers = "^4.14.0"
responses = "^0.26.0"

[tool.poetry.group.test.dependencies]
freezegun = "^1.2.2"
hypothesis = "^6.75.6"
pexpect = "^4.8.0"
pytest = "^8.2.0"
pytest-mock = "^3.14.0"
pytest-cov = "^4.0.0"
requests-mock = "^1.11.0"
nanoid = "^2.0.0"
testcontainers = "^4.14.0"

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | \.mypy_cache
  | \.pytest_cache
  | build
  | dist
)/
'''

[tool.isort]
profile = "google"
line_length = 120
known_first_party = ["strr_email"]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
minversion = "2.0"
testpaths = [
   "tests",
]
addopts = "--verbose --strict -p no:warnings --cov=src --cov-report html:htmlcov --cov-report xml:coverage.xml"
python_files = [
   "test*.py"
]
norecursedirs = [
   ".git", ".tox", "venv*", "requirements*", "build",
]
log_cli = true
log_cli_level = "1"
filterwarnings = [
   "ignore::UserWarning"
]
markers = [
   "slow",
   "serial",
   "conf: Inject flask config values into the app fixture",
]